Clean up hor
