+++ packages/block-library/src/footnotes/index.js
@@ -3,6 +3,8 @@
  */
 import { group as icon } from '@wordpress/icons';
 import { registerFormatType } from '@wordpress/rich-text';
+import { addFilter } from '@wordpress/hooks';
+import { privateApis as blockEditorPrivateApis } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
@@ -12,6 +14,8 @@ import edit from './edit';
 import metadata from './block.json';
 import save from './save';
 import { format } from './format';
+import { unlock } from '../private-apis';
+import hor from './higher-order-reducer';
 
 const { name } = metadata;
 
@@ -23,8 +27,11 @@ export const settings = {
 	save,
 };
 
-// Would be good to also remove the format if the block is unregistered.
+const { horKey } = unlock( blockEditorPrivateApis );
+
+// Would be good to remove the format and HoR if the block is unregistered.
 registerFormatType( 'core/footnote', format );
+addFilter( horKey, name, hor );
 
 export const init = () => {
 	initBlock( { name, metadata, settings } );
